Fixed a few warnings.
Rev: src/modules/CommonLog/clf.c:1.4
Rev: src/modules/HTTPLoop/accept_and_parse.c:1.18
Rev: src/modules/Image/encodings/gif.c:1.60
Rev: src/modules/Image/font.c:1.65
Rev: src/modules/Pipe/pipe.c:1.42
Rev: src/modules/files/file.c:1.189
Rev: src/modules/files/file.h:1.17
Rev: src/modules/files/sendfile.c:1.43
Rev: src/modules/system/nt.c:1.23
Rev: src/signal_handler.c:1.178

+++ src/modules/HTTPLoop/accept_and_parse.c
@@ -197,7 +197,7 @@ static int parse(struct args *arg)
   aap_get_header(arg, "content-length", H_INT, &arg->res.content_len);
   if((arg->res.data_len - arg->res.body_start) < arg->res.content_len)
   {
-    int nr;
+    ptrdiff_t nr;
     /* read the rest of the request.. OBS: This is done without any
      * timeout right now. It is relatively easy to add one, though.
      * The only problem is that this might be an upload of a _large_ file,
